completed Payment, add component AdminThemChuyenBay & AdminCreateFlight3M
