+++ admin.v2/index.html
@@ -5,6 +5,7 @@
     <link rel="icon" type="image/svg" href="./public/logo-header-word.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>TravFruit</title>
+    <script src="https://unpkg.com/react-scan/dist/auto.global.js"></script>
   </head>
   <body>
     <div id="root"></div>
