+++ admin.v2/package.json
@@ -24,6 +24,7 @@
     "react-hook-form": "^7.54.2",
     "react-paginate": "^8.3.0",
     "react-router-dom": "^7.1.1",
+    "react-scan": "^0.3.3",
     "react-toastify": "^11.0.3"
   },
   "devDependencies": {
