+++ admin.v2/src/Components/AdminLogin.jsx
@@ -27,7 +27,7 @@ function AdminLogin() {
   };
 
   const mutationLogin = useMutation({
-    mutationFn: Login, // Truyền function, không gọi trực tiếp
+    mutationFn: Login,
     onSuccess: (response) => {
       const { accessToken } = response.data;
       localStorage.setItem("accessToken", accessToken);
