+++ admin.v2/src/Components/CatchErrorAPI.jsx
@@ -1,7 +1,6 @@
 import PropTypes from "prop-types";
 
-function CatchErrorAPI({ error }) {
-
+function CatchErrorAPI({ error, handleAgain }) {
   return (
     <div className="bg-zinc-800 p-3 text-red-600">
       <p className="uppercase tracking-widest">Lỗi !!!</p>
@@ -17,10 +16,21 @@ function CatchErrorAPI({ error }) {
             "Không có error"}
         </p>
       </div>
+      {handleAgain && (
+        <button
+          className="text-white p-2 rounded-lg border border-red-600"
+          onClick={() => {
+            handleAgain();
+          }}
+        >
+          Thử lại
+        </button>
+      )}
     </div>
   );
 }
 CatchErrorAPI.propTypes = {
   error: PropTypes.object,
+  handleAgain: PropTypes.func,
 };
 export default CatchErrorAPI;
