+++ admin.v2/src/Components/CustomToast.js
@@ -1,0 +1,10 @@
+export const useToastOptions = {
+  position: "top-right",
+  autoClose: 5000,
+  hideProgressBar: false,
+  closeOnClick: false,
+  pauseOnHover: true,
+  draggable: true,
+  progress: undefined,
+  theme: "light",
+};
