+++ admin.v2/src/Components/convertDateToVN.js
@@ -1,0 +1,13 @@
+export const convertDateToVNDate = (dateString) => {
+  let vietnamDateTime = new Date(dateString).toLocaleString("vi-VN", {
+    timeZone: "Asia/Ho_Chi_Minh",
+    //weekday: "short",
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+    hour: "2-digit",
+    minute: "2-digit",
+    second: "2-digit",
+  });
+  return vietnamDateTime;
+};
