+++ admin.v2/src/main.jsx
@@ -1,3 +1,4 @@
+import { scan } from "react-scan";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router";
@@ -6,7 +7,9 @@ import "./index.css";
 import App from "./App.jsx";
 
 const queryClient = new QueryClient();
-
+scan({
+  enabled: true,
+});
 createRoot(document.getElementById("root")).render(
   <QueryClientProvider client={queryClient}>
     <StrictMode>
